async_api/requirements.txt
aioredis==2.0.0
elasticsearch[async]==8.6.0
fastapi==0.89.1
orjson==3.8.5
pydantic==1.10.4
uvicorn==0.20.0
uvloop==0.17.0
redis==4.4.2
fastapi-cache2==0.2.0
gunicorn==20.1.0
httpx==0.23.3
pyjwt==2.6.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-exporter-jaeger==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0